Fix full batch tests on Windows (#1704)
Several tests for "full batch" methods like GCN, RGCN and APPNP manually
construct the tensor inputs, in the form of a sparse adjacency matrix, specified
as a tensor of indices and a tensor of data. The indices are created as a Keras
`Input` with dtype `int64`, but the model was being passed NumPy arrays with
dtype `int32`. This was apparently fine on macOS and Linux, but not on Windows
(likely due to the LLP64 behaviour of that platform, see #1705 for more
discussion).

This PR updates the tests to pass in NumPy arrays with dtype `int64`, and also
updates the RGCN generator to do the same conversion internally (matching the
non-RGCN full batch generator).

See: #1699

+++ tests/layer/test_gcn.py
@@ -108,7 +108,9 @@ def test_GraphConvolution_sparse():
 
     # Note we add a batch dimension of 1 to model inputs
     adj = G.to_adjacency_matrix().tocoo()
-    A_indices = np.expand_dims(np.hstack((adj.row[:, None], adj.col[:, None])), 0)
+    A_indices = np.expand_dims(
+        np.hstack((adj.row[:, None], adj.col[:, None])).astype(np.int64), 0
+    )
     A_values = np.expand_dims(adj.data, 0)
     out_indices = np.array([[0, 1]], dtype="int32")
     x = features[None, :, :]
@@ -173,7 +175,9 @@ def test_GCN_apply_sparse():
     adj = G.to_adjacency_matrix()
     features, adj = GCN_Aadj_feats_op(features, adj)
     adj = adj.tocoo()
-    A_indices = np.expand_dims(np.hstack((adj.row[:, None], adj.col[:, None])), 0)
+    A_indices = np.expand_dims(
+        np.hstack((adj.row[:, None], adj.col[:, None])).astype(np.int64), 0
+    )
     A_values = np.expand_dims(adj.data, 0)
 
     generator = FullBatchNodeGenerator(G, sparse=True, method="gcn")
@@ -226,7 +230,9 @@ def test_GCN_linkmodel_apply_sparse():
     adj = G.to_adjacency_matrix()
     features, adj = GCN_Aadj_feats_op(features, adj)
     adj = adj.tocoo()
-    A_indices = np.expand_dims(np.hstack((adj.row[:, None], adj.col[:, None])), 0)
+    A_indices = np.expand_dims(
+        np.hstack((adj.row[:, None], adj.col[:, None])).astype(np.int64), 0
+    )
     A_values = np.expand_dims(adj.data, 0)
 
     generator = FullBatchLinkGenerator(G, sparse=True, method="gcn")
